requirements.txt
Flask==1.1.2
gunicorn==20.0.4
pymongo==3.10.1
Flask-Cors==3.0.8
Werkzeug==0.16.0
dnspython==2.0.0
